Explanation for Pet Store files
Store project
The goal was to create a database for a store with a database diagram as well. 
It should have at least 5 tables each with at least 3 columns. They will need to include at least 3 primary keys, and 2 foreign keys. There must be at least 9 different constraints on the columns, with at least 5 different types of constraints. There also must be 5 types of data.