SQL Dogs Only Vet Practice
I have created a database to house files with data related to patient care (details related to treating dogs), billing (details related to owners and separately details related to supplies for the practice), and management (staff and supplies). 
The data files were created using an online mock data generator, they were purposely created with datatype issues that would need to be addressed when loaded into the database.
The ETL (extract transform load) package that I created was designed to fix some datatype errors while loading the data files directly into the Dogs_only database. The concept is that a person could routinely use the same ETL package to upload newly created data files as often as necessary (hopefully, at the close of business daily) in order to allow staff with access to the Dogs_only database the ability to run reports adhoc and or/create new ones as needed. 
The SQL (a query language for data) relating to rabies shots that I created is for retrieving time-sensitive data that the practice would need often (realistically, continually) so much that it would be a waste of time to recreate the SQL for the information. I also created SQL to track the number of puppies (defined as patients aged 3 years or younger) within the practice. Finally, I created SQL to reverse all the actions of all SQL used on the database, this would restore the data to the state it was when it was loaded into the database.